requirements.txt
certifi==2023.5.7
pip==22.3.1
python-dateutil==2.8.2
setuptools==65.5.1
sib-api-v3-sdk==7.6.0
six==1.16.0
urllib3==2.0.3
wheel==0.38.4
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
